var fixes and addit
