+++ doc/TODO.md
@@ -3,5 +3,4 @@
 * Unit tests
 * Incorporate into express, similar to Jade, so compiling is done by the server
 * Editor plugin with syntax highlighting
-* File-import support
-* eval support
+* Asynchronous `parse` for [index.js](/index.js)
