+++ examples/demo_eval.jsml
@@ -1,0 +1,7 @@
+[
+    "The sum of 3 and 5 is: ",
+    {
+        "tag" : "b",
+        "eval" : "3 + 5"
+    }
+]
